+++ JavaAnalyzer/src/tool/compiler/java/ast/EquGenFieldExt.java
@@ -1,5 +1,6 @@
 package tool.compiler.java.ast;
 
+import polyglot.ast.Expr;
 import polyglot.ast.Field;
 import polyglot.ast.Node;
 import polyglot.ext.jl5.types.JL5FieldInstance;
@@ -32,12 +33,12 @@ public class EquGenFieldExt extends EquGenExprExt {
 	@Override
 	public Node equGenLeave(EquGenerator v) {
 		ReportUtil.leaveReport(this);
-		Field fld = (Field)this.node();
-		JL5FieldInstance fldIns = (JL5FieldInstance) fld.fieldInstance();
+		Field f = (Field)this.node();
+		JL5FieldInstance fldIns = (JL5FieldInstance) f.fieldInstance();
 		
 		// e.f / C.f
 		//   1. f의 타입 D를 가지오고, 새로운 집합변수 Chi2를 생성해서 D{Chi2}를 만들기
-		MetaSetVariable dchi2 = MetaSetVariable.create(fld.type());
+		MetaSetVariable dchi2 = MetaSetVariable.create(f.type());
 //		MetaSetVariable dchi2 = new MetaSetVariable(fld.type());
 		ReportUtil.report(dchi2, MetaSetVarSource.New, MetaSetVarGoal.Return);
 //		MetaSetVariable dchi2 = v.getCurrCC().getField(fldIns);
@@ -46,7 +47,7 @@ public class EquGenFieldExt extends EquGenExprExt {
 		//   2A. e.f	(non-static)
 		//       e의 타입 C{Chi1}을 가져오고, C{Chi1}.f <: D{Chi2}를 제약식 집합에 추가
 		if(!fldIns.flags().isStatic()) {
-			MetaSetVariable cchi1 = EquGenExt.metaSetVar(fld.target());
+			MetaSetVariable cchi1 = EquGenExprExt.metaSetVar((Expr)f.target());
 			ReportUtil.report(cchi1, MetaSetVarSource.Receiver, MetaSetVarGoal.Flow);
 			
 			ReadField rf = new ReadField(cchi1, fldIns, dchi2);
